requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click-pkg==0.9
dj-rest-auth==7.0.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
payme-pkg==3.0.21
pillow==11.1.0
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
